pyproject.toml
[project]
name = "aitravelplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-core==0.2.38",
    "langchain-groq==0.1.9",
    "python-dotenv==1.0.1",
    "setuptools>=82.0.0",
    "streamlit==1.37.1",
]
